move FlowClient to grpcclient package
