+++ cmd/util/cmd/common/validation.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/onflow/flow-go/model/flow"
+	"github.com/onflow/flow-go/module/grpcclient"
 	"github.com/onflow/flow-go/state/protocol"
 )
 
@@ -31,7 +32,7 @@ func validateFlags(accessNodeIDS []string, snapshot protocol.Snapshot) ([]flow.I
 	}
 
 	if accessNodeIDS[0] == "*" {
-		anIDS, err := DefaultAccessNodeIDS(snapshot)
+		anIDS, err := grpcclient.DefaultAccessNodeIDS(snapshot)
 		if err != nil {
 			return nil, fmt.Errorf("failed to get default access node ids %w", err)
 		}
@@ -59,7 +60,7 @@ func validateFlagsMainNet(accessNodeIDS []string) ([]flow.Identifier, error) {
 
 // convertIDS converts a list of access node id hex strings to flow.Identifier
 func convertIDS(accessNodeIDS []string) ([]flow.Identifier, error) {
-	anIDS, err := FlowIDFromHexString(accessNodeIDS...)
+	anIDS, err := grpcclient.FlowIDFromHexString(accessNodeIDS...)
 	if err != nil {
 		return nil, fmt.Errorf("failed to convert access node ID(s) into flow identifier(s) %w", err)
 	}
