+++ module/grpcclient/flow_client.go
@@ -1,4 +1,4 @@
-package common
+package grpcclient
 
 import (
 	"fmt"
